Use portable expression evaluation syntax
The 'region' command uses the two square brace evaluator syntax
"[[ expression ]]", which is a bashism and only works on platforms
such as OSX where /bin/sh is aliased to bash.

However, the single brace version '[ expression ]' works in all bourne
shell variants so is more portable.

Also changed the comment to reflect actual behaviour; the function
displays a single item.